Refactor a bit
- still need to write tests for errors
