+++ src/pepper/parser.py
@@ -19,7 +19,7 @@ import pepper.abstract_symbol_tree as ast
 from pepper.lexer import lexer
 from pepper.lexer import tokens  # NOQA
 import pepper.symbol_table as symtable
-
+from pepper.evaluator import parse_macro
 # precedence according to
 # http://en.cppreference.com/w/c/language/operator_precedence
 
@@ -695,254 +695,6 @@ def p_error(p):
     raise symtable.PepperSyntaxError()
 
 
-def validate(exp, line_no):
-    literal = None
-    if isinstance(exp, ast.PreprocessingNumberNode):
-        if not exp.children[0].isdigit():
-            print(f"ERROR: token {exp.children[0]} is not valid in preprocessor expressions")
-            raise symtable.PepperSyntaxError()
-        literal = exp.children[0]
-    elif isinstance(exp , ast.StringLiteralNode):
-        if  len(exp.children[0]) != 3:
-            print(f"ERROR: token {exp.children[0]} is not valid in preprocessor expressions")
-            raise symtable.PepperSyntaxError()
-        literal = str(ord(exp.children[0][1]))
-    elif isinstance(exp, ast.ASCIILiteralNode):
-        literal = exp.children[0]
-    elif isinstance(exp, ast.OperatorNode):
-        child = exp.children[0]
-        if child == '&&':
-            literal = 'and'
-        elif child == '||':
-            literal = 'or'
-        else:
-            literal = child
-    elif isinstance(exp, int) or isinstance(exp, bool):
-        literal = str(exp)
-
-
-    return literal
-
-
-def parse_line(t):
-    if isinstance(t[0] , ast.LinesNode):
-        return parse_lines(t[0])
-    while isinstance(t[0], ast.IdentifierNode):
-        if t[0].children[0] in symtable.TABLE:
-            t = symtable.TABLE[t[0].children[0]].tokens
-        else:
-            t = [0]
-
-    return t
-
-
-def parse_lines(token):
-    return [parse_line([t]) for t in token.children]
-
-def parse_parant(ev, i, beg):
-    temp = []
-    keep_going = 1
-    if beg:
-        temp.append(ev[i])
-        i +=1
-        while keep_going:
-            if ev[i] == '(' :
-                keep_going += 1
-            elif ev[i] == ')':
-                keep_going -= 1
-
-            temp.append(ev[i])
-            i += 1
-
-    else:
-        temp.append(ev.pop())
-        while keep_going:
-            if ev[-1] == '(' :
-                keep_going -= 1
-            elif ev[-1] == ')':
-                keep_going += 1
-
-            temp.append(ev.pop())
-        temp = temp[::-1]
-
-
-
-    return temp, i
-
-def unravel_list(exp):
-    parsing = []
-    is_white = True
-    for tok in exp:
-        while isinstance(tok, list):
-            temp = tok.pop(0)
-            if not tok and not isinstance(temp, list):
-                tok = temp
-            elif not tok:
-                parsing.append(ast.ASCIILiteralNode(['(']))
-                tok = temp
-                parsing.append(ast.ASCIILiteralNode([')']))
-            elif not isinstance(temp, list):
-                parsing.append(temp)
-            else:
-                parsing.append(ast.ASCIILiteralNode(['(']))
-                parsing.extend(unravel_list(temp))
-                parsing.append(ast.ASCIILiteralNode([')']))
-
-
-
-        parsing.append(tok)
-
-    i = 0
-    while i + 2 < len(parsing):
-        found = False
-        if isinstance(parsing[i], ast.ASCIILiteralNode) \
-and isinstance(parsing[i+1], ast.WhiteSpaceNode) \
-and isinstance(parsing[i+2], ast.ASCIILiteralNode):
-            parsing.pop(i)
-            parsing.pop(i)
-            parsing.pop(i)
-            found = True
-        if not found:
-            i +=1
-
-    return parsing #[ast.ASCIILiteralNode(['('])] + parsing + [ast.ASCIILiteralNode([')'])]
-
-
-def parse_macro(tokens, line_no = 0 ):
-    scalar_tokens = []
-    for token in tokens:
-        token = [token]
-        if isinstance(token[0], ast.LinesNode):
-            token = parse_lines(token[0] )
-        while isinstance(token[0], ast.IdentifierNode):
-            if token[0].children[0] in symtable.TABLE:
-                token = symtable.TABLE[token[0].children[0]].tokens
-            else:
-                token = [0]
-
-
-        scalar_tokens.append(token)
-
-    evaluation = []
-    # no defined macros will accepted if not an integer type
-    for expr in scalar_tokens:
-            for exp in expr:
-                if isinstance(exp, list):
-                    exp = unravel_list(exp)
-                    exp = [validate(tok, line_no) for tok in exp if not isinstance(tok, ast.WhiteSpaceNode)]
-                    evaluation.extend(exp)
-                elif exp == "(" or exp == ")":
-                    evaluation.append(exp)
-                elif not isinstance(exp, ast.WhiteSpaceNode):
-                    exp = validate(exp, line_no)
-                    evaluation.append(exp)
-
-
-
-    OPERATORS = {'+', '-', '*', '/', 'and', 'or', '&', '|', '<<', '>>', '^'}
-    i = 0
-
-
-    while i + 1 < len(evaluation):
-        found = False
-        if evaluation[i] == "("  and  evaluation[i+1] == ")" :
-            evaluation.pop(i)
-            evaluation.pop(i)
-            found = True
-        if not found:
-            i += 1
-
-    i = 0
-
-    while i + 2 < len(evaluation):
-        found = False
-        if evaluation[i] == '(' and evaluation[i+1] in OPERATORS and evaluation[i+2] == ')':
-            evaluation.pop(i)
-            evaluation.pop(i+1)
-        i +=1
-
-
-
-    print(evaluation)
-    # catch AND  && OR stuff
-    i = 0
-    new = []
-    while i < len(evaluation):
-        temp= []
-        token = evaluation[i]
-        if token == 'and':
-            temp.insert(0 , 'bool(')
-
-            if evaluation[i-1] != ')' :
-                temp.append(new.pop())
-            else:
-                left, i = parse_parant(new, i , False)
-                temp.extend(left)
-
-            temp.append('and')
-
-            if evaluation[i + 1] != '(':
-                temp.append(evaluation[i+1])
-                i +=1
-            else:
-                right, i = parse_parant(evaluation, i+1, True)
-                temp.extend(right)
-                i +=1
-
-            temp.append( ')')
-            new.extend(temp)
-
-        elif token == 'or':
-            temp.insert(0 , 'bool(')
-
-            if evaluation[i-1] != ')' :
-                temp.append(new.pop())
-            else:
-                left, i = parse_parant(new, i , False)
-                temp.extend(left)
-            temp.append('or')
-
-            if evaluation[i + 1] != '(':
-                temp.append(evaluation[i+1])
-                i +=1
-            else:
-                right, i = parse_parant(evaluation, i+1, True)
-                temp.extend(right)
-                i +=1
-            temp.append( ')')
-            new.extend(temp)
-
-
-
-        else:
-            new.append(token)
-
-
-        i +=1
-
-    evaluation = new
-
-
-
-
-    print("\n",evaluation)
-    try:
-        final = eval(" ".join(evaluation))
-    except SyntaxError:
-        print(f"ERROR: syntax error")
-        raise symtable.PepperSyntaxError()
-
-
-    if (isinstance(final, str) and len(final) > 1) and not isinstance(final, float) and not isinstance(final , int):
-        print(f"ERROR: token {final} is not valid in preprocessor expressions")
-        raise symtable.PepperSyntaxError()
-
-    if isinstance(final, float):
-        final = int(final)
-
-
-    print ( final )
-    return final
 
 def get_args():
     parser = argparse.ArgumentParser()
